View breadth requirements for UTSC courses
This change will allow for courses that are at UTSC to have breadth requirements. Previously the scraper would only fetch breadths for UTSG courses. I have changed that now so we can populate breadths for UTSC courses as well. It still holds the model validations so "breadths" is still a list of numbers. 

For UTSC, the numbers correspond the breadth in this fashion:
1 => Arts, Literature & Language
2 => "History, Philosophy & Cultural Studies"
3 => "Natural Sciences"
4 => "Social & Behavioural Sciences"
5 => "Quantitative Reasoning"

UTSC courses don't have more than one breadth so I feel this change should work for all courses at UTSC. I have tested it and confirmed the outputted data to be correct.